web/requirements.txt
mysqlclient==2.0.3
django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
Pillow==8.3.1
drf-yasg==1.20.0
django-allauth==0.45.0
firebase-admin==5.1.0
psycopg2>=2.8